+++ src/components/Input/CardNameInput/CardNameInput.styles.ts
@@ -1,0 +1,15 @@
+import styled from 'styled-components';
+
+export const Styled = {
+  CardNameInputWrapper: styled.div`
+    display: flex;
+    margin-top: 125px;
+    input {
+      width: 240px;
+      text-align: center;
+      border: none;
+      font-size: 18px;
+      border-bottom: 1px solid black;
+    }
+  `,
+};
