+++ src/components/Input/CardNumberInput/CardNumberInputs.styles.ts
@@ -1,0 +1,13 @@
+import { Container } from 'components/style/InputContainer';
+import styled from 'styled-components';
+
+export const Styled = {
+  CardNumberContainer: styled(Container)`
+    justify-content: center;
+    width: 318px;
+    Input {
+      width: 15%;
+      text-align: center;
+    }
+  `,
+};
