+++ src/components/Input/ExpirationDateInput/ExpirationDateInput.tsx
@@ -1,8 +1,9 @@
 import { Input } from 'components/common';
 import React, { ChangeEventHandler, ChangeEvent } from 'react';
-import styled from 'styled-components';
-import { ValueAndOnChange } from './types';
+import { ValueAndOnChange } from '../types';
 import { validateInput } from 'util/Validation';
+import { Styled as S } from './ExpirationDateInput.styles';
+import FormLabel from 'components/common/FormLabel/FormLabel';
 
 export interface ExpirationProps {
   month: ValueAndOnChange;
@@ -30,33 +31,30 @@ export function ExpirationDateInput(props: ExpirationProps) {
 
   return (
     <>
-      <Input
-        ref={inputRefs[0]}
-        inputMode="numeric"
-        value={month.value}
-        type="text"
-        maxLength={2}
-        placeholder="MM"
-        onChange={(e) => handleChange(e, 0, month.onChange)}
-        required
-      />
-      <SLASH />
-      <Input
-        ref={inputRefs[1]}
-        inputMode="numeric"
-        value={year.value}
-        type="text"
-        maxLength={2}
-        placeholder="YY"
-        onChange={(e) => handleChange(e, 1, year.onChange)}
-        required
-      />
+      <FormLabel>만료일</FormLabel>
+      <S.ExpirationDateContainer>
+        <Input
+          ref={inputRefs[0]}
+          inputMode="numeric"
+          value={month.value}
+          type="text"
+          maxLength={2}
+          placeholder="MM"
+          onChange={(e) => handleChange(e, 0, month.onChange)}
+          required
+        />
+        <S.SLASH />
+        <Input
+          ref={inputRefs[1]}
+          inputMode="numeric"
+          value={year.value}
+          type="text"
+          maxLength={2}
+          placeholder="YY"
+          onChange={(e) => handleChange(e, 1, year.onChange)}
+          required
+        />
+      </S.ExpirationDateContainer>
     </>
   );
 }
-
-const SLASH = styled.span`
-  &::before {
-    content: '/';
-  }
-`;
