--- src/components/Input/OwnerNameInput.stories.tsx
@@ -1,12 +1,0 @@
-import { StoryFn } from '@storybook/react';
-import { OwnerNameInput, OwnerNameInputProps } from './OwnerNameInput';
-
-export default {
-  title: 'OwnerNameInput',
-  component: OwnerNameInput,
-};
-
-const Template: StoryFn<OwnerNameInputProps> = (args) => <OwnerNameInput {...args} />;
-
-export const Default = Template.bind({});
-Default.args = {};
