--- src/components/Input/OwnerNameInput.tsx
@@ -1,19 +1,0 @@
-import { Input } from 'components/common';
-import { ChangeEventHandler } from 'react';
-
-export interface OwnerNameInputProps {
-  value: string;
-  onChange: ChangeEventHandler<HTMLInputElement>;
-}
-
-export function OwnerNameInput({ value, onChange }: OwnerNameInputProps) {
-  return (
-    <Input
-      value={value}
-      type="text"
-      maxLength={30}
-      placeholder="카드에 표시된 이름과 동일하게 입력하세요"
-      onChange={onChange}
-    />
-  );
-}
