+++ src/components/Input/index.js
@@ -1,5 +1,5 @@
-export * from './CardNumberInputs';
-export * from './ExpirationDateInput';
-export * from './OwnerNameInput';
-export * from './PasswordInput';
-export * from './SecurityCodeInput';
+export * from './CardNumberInput/CardNumberInputs';
+export * from './ExpirationDateInput/ExpirationDateInput';
+export * from './OwnerNameInput/OwnerNameInput';
+export * from './PasswordInput/PasswordInput';
+export * from './SecurityCodeInput/SecurityCodeInput';
